Fix make distcheck javascript inculde path
gjs will look for [searchpath]/overrides/[module name] when importing
a new typelib module. We had $(top_builddir) as part of the include
path, but not $(top_srcdir). As our override file is no longer
generated by make, it is in $(top_srcdir)/overrides and not
$(top_builddir)/overrides.

Added $(top_srcdir) to the path, fixing make distcheck which does
and out of srctree build.
[endlessm/eos-sdk#826]