Workaround for bug with Gtk-Doc 1.20.1
With (at least) version 1.20.1 of Gtk-Doc, 'make check' would fail
because GTKDOC_CHECK was given the value gtkdoc-check.test instead of
the path to the program. I'm not sure if that is a bug or a feature,
but using GTKDOC_CHECK_PATH doesn't work with 1.18, so this adapts for
both versions.

Also fixed a typo.

[endlessm/eos-sdk#833]