LessonCard widget
LessonCard widget in Javascript override. It represents a how-to
article, or a lesson from a Khan Academy course. It can be given an
integer index which might be displayed as something like "Lesson 1",
and it can be set to "complete", which might be displayed as a
checkmark or something similar.

When it is set to complete, it gets the style class "complete".

[endlessm/eos-sdk#833]