Progress card
This Card subclass represents a group of cards in a sequence that can
be completed or "checked off". Depending on how the display is
implemented, 5 total-items and 4 completed-items could be displayed as
"4 out of 5 finished" or "80% done" or something entirely graphical.
The current manner of displaying them is a placeholder, pending design
specs.

[endlessm/eos-sdk#836]