Call Gtk.init before running gtk tests
We had been making a number of gtk calls as part of our tests, but
had not been initializing gtk before hand. This is why our tests
when they failed would contain huge dumps of assertion fails.

In the case of a GtkScrolledWindow which is coming in the section
view, the tests will just segfault. We need to call Gtk.init before
running any gtk code in out tests
[endlessm/eos-sdk#977]