Article page widget for template A
Properties are pretty basic, they just expose the API of the table
of contents and the switcher widget.

Most of the complexity is in the sizing. Was hoping to just use a grid
but that gave an unreasonable amount of space to the table of contents
because a webview can render at any size and has no size request. So
had to reach for custom container again, to enforce a rule that the
table of contents should be at most 25% the width of the total
allocation.

Added two internal frames for setting things like padding for the
toolbar on the right and around the webview. Otherwise the widget
was basically unthemeable.
[endlessm/eos-sdk#960]