Don't pass css matchers when matching undefined style class
Not sure why we had it in there as such. This just had me confused
for a while because matching a misspelled style class constant
(such as expect(widget).toHaveCssClass(EosKnowledge.DOES_NOT_EXISTS)
will pass unconditionally. Think it will be better to fail on undefined
to catch these kind of errors.
[endlessm/eos-sdk#960]