Home page A functions without cards added
The container for holding the cards wasn't working when no cards
were set. Now it should work without warnings.

Couldn't think of a good way to test this without mapping a window
to the screen, as the bugs were in size_allocate and
get_preferred... functions. Best way to test for now is comment out
adding cards in the home page smoke test. In the future this would
be a good place for dogtail.
[endlessm/eos-sdk#967]