transition_duration sets properly after construction
In lightbox and section article page. We had these bound to gtk
widget properties but they do not set properly after construction
without a getter and setter defined

To test we would need to reach into the internals of the widget
and assert that the stack and revealers properties update as they
should. But that would be testing the implementation, so didn't
put that in
[endlessm/eos-sdk#997]