Stop using overrides for javascript code install
Instead install EosKnowledge.js to a location on the GJS_PATH.
Instead of overrides just mess with the imports.gi.EosKnowledge
object directly in EosKnowledge.js. This gives us the rather ugly
path of imports.EosKnowledge.EosKnowledge, but makes the contents
of EosKnowledge a lot simpler.

And of course, exceptions in our javascript code just get reported
normally now
[endlessm/eos-sdk#1026]