Previewer widget with image support
The previewer widget and the first of its sub widgets the image
previewer. Since we are not sticking to the fancy plugin
architecture for the near term, just the kept the main Previewer
widget as simple as possible.

Most complexity is from the imagePreviewer which sizes the image
down the the available allocation unlike GtkImage. Thought that
was import in case we ever stick big images in our previewer, it
shouldn't constrain the window size.
[endlessm/eos-sdk#1015]