Styled the front page search bar
The designs gave a image as the search background, but imitated
with css so we can use stock gtk widgets. The search icon is a
little too small, but to fix it we'd need to move from
GtkSearchEntry to GtkEntry, which mean we would loose some of the
nice features of GtkSearchEntry. So leaving it for now

[endlessm/eos-sdk#1289]