Presenter shows media objects in lightbox
The presenter handles the events when a media object is clicked in an article page to be displayed in a lightbox. Clicking on the navigation buttons on the lightbox will display the corresponding media objects according to the position in the resources array of the article model.

The ArticlePresenter decides whether the clicked object is a local media object, external media object (i.e. does not belong to an article object) or a new page to be loaded.

[endlessm/eos-sdk#1070]