Updated tab button css for hover and active
Updated the css backgrounds for the tab button, and added the
`hover` and `active` states. Also changed css colors to (roughly)
match the mockups, and used `@define-color` to make the colors
more reusable.
Note that the gradient shades are a bit of guesswork based on the
mockups, not exact numbers from the design team.
[endlessm/eos-sdk#1329]