Home page switches based on template
Window now contains a template-type property, it will switch the
hope page it uses based on that property value.

Presenter reads the template-type from the app.json and will add
the correct type of cards to the home page depending on the
template.

Had to switch the Window to be constructed in the presenter, so
Window could have a construct only template-type property. But
think that's a more consistent switch. Presenter should set up
all the views and models
[endlessm/eos-sdk#1421]