Loosen check for card number in template B home page
The template only officially support 4, 6 and 8 categories and
thats the number we should test at and focus our efforts on. However
given how varied on content can be, I think it will be better to
just printerr a warning but not crash immediately.

That will make us more robust to our content, which I think is a
good thing.
[endlessm/eos-sdk#1421]