Reworked the virtual functions for widget layout
In card and home page, before we were relying on some subset
of the supclass member data being public and used by the subclasses.
Switched to the virtual functions passing the widgets that need to
be packed.

Making the vfunc role more explicit makes more sense I think and
will lead to a more consistent API
[endlessm/eos-sdk#1421]