Wait for webview switcher transition
In GTK 3.12 we have the 'transition-running' property on Gtk.Stack, so
we can do certain actions after the animations for a new page have
finished.

Since the first displayed page does not have an animation, this adds a
'display-ready' signal to the webview switcher that is emitted either
when the first displayed page is done loading, or when the animation
for a subsequently displayed page has completed.

[endlessm/eos-sdk#1528]