Mock out webview in article presenter test
The real webview was somehow making requests to the knowledge engine
(a URI in the testing code?) These were failing on Jenkins because
there is no knowledge engine running there.

The Engine class (and, in fact, the view) should be mocked out of this
test, but I'll open a new issue for that.

[endlessm/eos-sdk#1528]