Fix scrolling by explicitly selecting IDs in smooth_scroll
Previously, we were using document.querySelector() to select
anchors by ID in the page to which we wanted to scroll. Wikipedia
often has periods in their ID names, so querySelector thinks you
are trying to select a class. Now we explicitly select by
[id="name_of_id"] so that it looks for a tag with that literal id
and doesn't mistake periods for the beginning of a class selector.