Show top title label when no table of contents exists
Previously, when we had an article with no table of contents,
we wouldn't have a visible title because the side title got
hidden with the toc. In this commit we make the binding of the
top title bidirectional with the binding of the side title
so that whenever one is hidden, the other is visible.