Redo lightbox sizing
It was broken for large image in a number of ways. Switch the way
the lightbox container works, now it is just one widget with some
dynamic style classes applied.

All the complexity is now in the size_allocate routine, which is
fairly huge, but should have the lightbox sizing properly no
matter the image or window size
[endlessm/eos-sdk#1653]