Mark strings for translation
The "Articles" section page segment label and the "Results for ..."
search results label were not marked for translation.

Strictly speaking translatable strings are a UI thing and should
therefore be part of the view, not the presenter, but we'll have to
change that in a later code audit -- we're in a hurry.

[endlessm/eos-sdk#1711]