Added infinite scrollbar.
When the scrollbar hits the bottom of the page, or when
the page hasn't been filled, the scrollbar emits a signal
to lazy load additional content.

Known Issue: The lazy loading re-requests all the existing
content in addition to the new content. This should change
to only loading the new content when Xapian can handle
chunked requests.
--Rob&Ror