Table of contents animates to target-section
To fix up our scrolling animations, we need to get the table of
contents to power its own animations rather than follow the
current anchor in the webview. So this commit adds a target-section
property which the selected-section will constantly animate towards.

Also removed an old feature of the table of contents, where -1 could
be set as a selected-section to remove the highlight from any entry
in the table of contents. We were never using the feature, and it was
going to add a lot of complexity, so out it goes.
[endlessm/eos-sdk#1705]