Redo smooth scrolling inside webview
Using jquery. The old smooth scrolling was using some standalone
script some dude had written, but it didn't always transition in
exactly the duration specified, which made it impossible to sync
with an animation from Gtk.

jquery, though its a enormous bag of tools, allows us to drop 75%
of the scrolling code, so less maintenance for us, and it works
with proper timing.
[endlessm/eos-sdk#1705]