Better error checking in engine requests
A proper fix of the one I botched earlier. We should
be checking if the response from the EKN is null during
marshalling, NOT during the _send_json_ld method, since it
is in fact legitimate for it to return null. Also, in the
presenter, we should only try to reference results from the
knowledge engine if error is not defined.