Buttons queue resize on state flags changed
Our back button and template A cards change size on hover state
via changes to margin and padding in css. These could occasionally
get stuck at the hover size when the mouse was not over then. Possibly
a bug in GTK? Rigging up the state flags change signal to a queue_resize
fixes the issue
[endlessm/eos-sdk#1674]