See all categories hides itself when not necessary
Added an all-cards-visible property to the home page a container
and used that to update the button visibility state.

Then ran into problems where updating the button visible would
automatically trigger the category page transition, so needed
to change how that was handled as well.

Switched to a model where the toplevel window just lets the pages
know if they are currently animating, and they handle showing and
hiding the buttons themselves
[endlessm/eos-sdk#1846]