Always load_uri via the webview switcher
Previously we were only loading a URI through the webview
switcher if we wanted an animation of the webpage - otherwise
we would just load it in the static webview. However, that would
have meant duplicated code when, for example, we want to always
grab focus on the webview after loading. This does a little
refactoring so that we can always load via the switcher, and
simply return false from the decide-navigation-policy when
we don't want an animation.

[endlessm/eos-sdk#1840]