Only add new history model item if loading new page
Previously, we would always add a new history item when
we went to another page, even if that page's URL was the
same as the one we were currently on. This doesn't make
sense so instead now we only add a new history model
item if we are navigating to a new page.

[endlessm/eos-sdk#1885]