No-search-results label is ellipsized properly
Fixed a label that was being truncated too short when needed to be ellipsized. Unfortunately this revealed another bug with the placement of the other two labels in NoSearchResultPageB. After a lot of head-scratching, we concluded that the best solution (for the little gain of this page, that is)  is to have different widgets for templates A & B, while still reusing the same translations.

[endlessm/eos-sdk#1880]