contentUrl is now part of ContentObjectModel
To serve PDF documents, we are adding a contentUrl property to the ContentObjectModel, so that the ArticleObjectModel can use it to point to the path of the PDF file. Since this property was already being used by the MediaObjectModel, we are refactoring these objects to move the property up in the class hierarchy.

[endlessm/eos-sdk#1355]