Refactor common code
The code for grabbing the focus and emitting the display-ready signal
should go in a helper function because it is already (sort of) used in
two places and the current issue requires it to be used in a third
place as well.

Likewise, load-changed should be disconnected right after the
load-changed handler decides it doesn't need to be connected any more;
another code path needs to be added in the current issue, and we can
disconnect that signal once before either code path occurs.

(Kurt & Philip)

[endlessm/eos-sdk#1352]