WebviewSwitcherView::create-view-for-file signal
This replaces the ::create-webview signal with ::create-view-for-file
so that other documents can be shown besides webviews: for example,
an embedded Evince PDF viewer.

This does require that the widget returned from the signal handler
implements part of the WebKit2.WebView interface (see the documentation
for more details.) This was already the case with ::create-webview, so
we already have one adapter class: MockWebView, used in the tests.
Another one for EvinceView will be added in a subsequent commit.

In addition, this does remove the ability to reuse the same webview if
loading a new URI with animation type set to NONE. Since it is up to
the client code to determine what view to use for what URI, this code
has no way of knowing whether the client will want to use the same view
or not.

(Kurt & Philip)

[endlessm/eos-sdk#1352]