Removed switcher api, just use a stack in articlePage
This removes a lot of code overhead to drive the loading of
article content entirely from within the view. Instead handle
loading of new views from within the article presenter, and
only handle transition the views within the view.

With this change the webviewSwitcherView and evinceWebviewAdapter
go away entirely, the switcher becomes a stock GtkStack within the
article page.

[endlessm/eos-sdk#2069]