Shared library's init needs XDG_RUNTIME_DIR
The init code of the shared library, which is called every time the
shared library is loaded -- including for GObject introspection and
Gtk-doc scanners -- tries to initialize clutter. That will fail, but
it will also abort the program entirely if XDG_RUNTIME_DIR is not set.
Therefore, we set it to /tmp as a workaround.