Table of contents stays constant sized on the article page
We have specs for what fraction of the article view the toc
should take up. However, we were using min of that fraction
and the tocs natural size to come up with the actual allocation.
So tocs with small entry labels (and a small natural size) would
size below the desired fraction of the article view.

Its not incorrect to be handing the toc a size bigger than its
natural size. And to keep the table of contents at a constant size
we should be

[endlessm/eos-sdk#2040]