Template A card thumbnails display at prober size
Their height was wrong, so now hard coded the height in the css.
It turned out to be remarkably tricky to get things to animate
and render properly. I tried hard coding the width as well in css,
but the image would jitter oddly when animating to its larger size.
So I left the image hexpanding to kill the jitter.

Finally even when the image is rendering at the right size, it will
still be fuzzed when background-size: cover; is used. So using
background-size: 100% 100%; This gives us clarity of those images,
but will break the aspect ratio of image that do not conform to
the 163x133 size
[endlessm/eos-sdk#2044]