Ensure callbacks are only called once in Engine.js
These changes ensure that the Engine methods return immediately after
calling their asynchronous callbacks, and that callbacks are not called
within a try block -- for example, if a successful callback was called
within a try block and threw an exception, that exception would cause
the error callback to be called as well.

This adds tests for the bug, but the tests themselves do not fail;
instead, the presence of the test causes other tests to fail without
this fix! That would indicate that callbacks are being called more than
they ought to be, and generating exceptions during other tests.

(Fernando, Patrick, Philip)

[endlessm/eos-sdk#2131]