articlePresenter test always waits for test article content
In our article presenter, we ask it to load up a new page of
content. The presenter will hand a new page to the view, and
only updates it toc and title when the view emits
new-view-transitioned.

We used to be running our tests as soon as the presenter handed
off the new article page to the view. And relying on some stack
calls and signal handlers to finish before the tests actually
started. Which appears to only hold true in some versions of GLib.
A more robust approach is to only run the tests after
new-view-transitioned is called.
[endlessm/eos-sdk#2230]