Only set search box query on user entered search
Previously we were clearing the top-bar search box
when the user selected an article and then
programmatically repopulating it when the user
went back the search page. But this makes search
entry think it should fire off another autocomplete
request. So instead we leave the search bar to have
what the user put in.

[endlessm/eos-sdk#2306]