Search box only suggests prefixed titles.
Since we index articles by both their original (wikipedia) title and
the title assigned by content, when users received autocomplete
suggestions in the search box, they always showed the article's title,
even when it did not contain their search as a substring.

Now it will show either the original or assigned title, depending on
which one is prefixed by the user's query.

[endlessm/eos-sdk#2397]