Fully implement debug buttons
This connects up the debug buttons with all the actions they are supposed
to accomplish. You can go backwards to the previous issue or forwards to
the next issue. The "previous issue" button is insensitive if you are at
the first issue.

Extra care is required to release all references to webviews held by the
presenter when loading a new issue. Otherwise the webviews stick around
and cause 'cannot disconnect handler with ID ###' messages when the app
quits.

One limitation is that the "next issue" button does not turn insensitive
if you are at the last issue; you just get the error page instead.

[endlessm/eos-sdk#2487]