Make application->preseter API explicit
This was always been gross, the application has called
> this._presenter._on_search(this._presenter.view, query);
for a while, just reaching into private functionality of the
presenter.

But now that the application lives outside of the knowledge lib
library, I think making a solid API the application class can use
becomes even more important.

With this commit, the Application expects the presenter to provide two functions
search: function (query)
activate_search_result: function (query, model, results, more_results_fn)
for activating search results and activating a specific article.

Added stubs for these in the reader presenter to make it clear that search
results are not yet supported by those application, and the work that would
be needed to make search results possible.
[endlessm/eos-sdk#2467]