Rearrange widgets on article page
The specs show that the title view, in some styles, will have to extend
all the way to the top of the page, because it will have a border. It's
also centered vertically in the page, so having empty space on top of it
belonging to the progress label was making it appear too low on the page.

This rearranges the article page's grid slightly so that the widgets
appear in more appropriate places.

[endlessm/eos-sdk#2506]