Logic for updating the issue at regular intervals.
Currently set to update the issue every week. The _update_issue
function currently grabs the articles in the next issue, but
this can be changed if we decide to go with a more feed-like,
rolling version of 'issues'.The logic is in the presenter to
provide flexibillity in the case that we want to make several
different queries for the same issue (e.g. grab some articles
tagged with issue14 and some with issue15).

[endlessm/eos-sdk#2438]