More robust handling of webview policy logic
In preparation of the commit ahead of this one,
I tidied up the logic of handling webview
navigation requests. Specifically, we now
check for ekn_id matches using only the hash
suffix of a URI, rather than the full uri.
This ensures that we can change
localhost:3003/api/domain to something else
and still have correct id matches.

[endlessm/eos-sdk#2510]