Handle images with ekn:// scheme in eknWebView
We now make the data in the database agnostic about
the host and port of the knowledge engine. So all
URIs simply have the ekn:// scheme. This commit
ensures that knowledge lib translates those ekn://
URIs to the correct knowledge engine uri, i.e.
with the right host and port set. Image URIs are
handled with the custom registered scheme callback;
article URIs are handled in the decide-policy
callback.

[endlessm/eos-sdk#2510]