Patch engine to work with old databases
We will carry around a patch to continue to function with eos2 knowledge
bundles prior to 2.3. These databases had the host and port of the
old nodejs knowledge engine hardcoded all of the database.

If we no longer need to worry about these old bundles (e.g. eos3)
please revert.
[endlessm/eos-sdk#2608]