Search XDG data dirs for the EKN manifest
Instead of hardcoding /endless/share as the location to look for EKN
Xapian databases, this searches through the possible data dir paths
according to the XDG directory spec for the EKN database path.

It uses a fallback of /endless/share if that path isn't already in the
list of data directories.

[endlessm/eos-sdk#2635]