Add article templating with mustache
Porting over the article templating from jade in the knowledge engine.

Kept all the complex localization to the javascript, so we didn't
have to try to homebrew any complicated localization in templates
solutions.

For testing tried to keep things so that small changes to the template
html wouldn't break the tests. Focused on the logic in
articleHTMLRenderer
[endlessm/eos-sdk#2638]