Use the ekn:// article uri when loading article in webview
Before we were using "bogus-uri", but I think this makes more sense.
It's the actual uri of the article we are showing.

Also since this is a real uri scheme, we can register the ekn uri a
local uri for security purposes, meaning we can load resources from
the local file system.
[endlessm/eos-sdk#2638]