Add mathjax to the article template
We are using a debian packaged version of mathjax and including that.
So we very basically check for /usr/share/javascript/mathjax
in configure.ac (as there's no package config file) and allow
an optional configure parameter to point to a different location.

Added all the configuration grossness to the mustache template,
though it could easily live in the js side of things too.
[endlessm/eos-sdk#2638]