contentObjectModel: only reference ekn ids, not other models
Gets rid of all the complex code to handle deferred loading of
certain property in the model. In the new world order after this
commit, the json+ld for each model will always contain everything
needed to fully instantiate the model.

We used to do all the deferred loading to prefetch all the image
models we would need to render the page. With the new stateless
ekn:// uri handler, this wasn't happening anyway. And with the
knowledge engine out of the picture, queries to our database
take much less time.
[endlessm/eos-sdk#2700]