Makefile: Rename some test makefile variables
Because of all the makefile including we do, all our makefile variables
are competing the the same name space. Added a test_ prefix to
variables we define in the the test makefile to keep from having collisions.

We had a variable in tests called gresource_deps and another target in our
c lib makefile called gresourcedeps. It was making me uncomfortable.
[endlessm/eos-sdk#2638]