engine: fix tags in get_object_by_query
After https://github.com/endlessm/eos-sdk/issues/2636 our handling
of tags in engine.js got a little muddled.

We had a query parameter called tag which according to the doc
comments needed to be a string comma separated list of tags, but
in fact needed to be a list.

Templates A and B were calling into with a list, but the reader code
was calling into with a string.

Changed the docs to ask for a tags object which is a javascript list
and used the new API consistently across templates A/B/reader
[endlessm/eos-sdk#2638]