Don't give 64-bit ints to 32-bit properties
The maximum integer for 32-bit int properties should be GLib.MAXINT32 or
GLib.MAXUINT32, depending on whether it is signed or unsigned. When a
property needs to have 64 bits, such as a timestamp in ms, explicitly
make it a 64-bit int property.

Note that GLib.MAXUINT64 doesn't marshal properly into
GObject.param_spec_uint64's maximum argument slot, so we have to leave
the maximum as GLib.MAXINT64.

[endlessm/eos-sdk#2681]