Refactor current page into presenter
Since the view should be as dumb as possible, we refactor the Reader
window's current_page property into the presenter. This means the
presenter keeps track of the data it needs, and which the view doesn't
have to know about.

Instead of setting the view's current_page property to change to another
page, the view now has show_article_page(), show_overview_page(), and
show_done_page() methods. In show_article_page() you have to specify a
transition direction, since the view doesn't necessarily know which
direction to go now that we have nonlinear navigation. For now, we always
transition forward to the done page and backward to the overview page.

(Originally by Rory; commit split out, code review changes made, and
tests added by Philip)

[endlessm/eos-sdk#2681]