Launcher interface for presenters
There are three code paths with which a knowledge app can be started up:
straight from the desktop or command line, with a search query, and with
instructions to display a particular search result. This introduces an
interface, EosKnowledge.Launcher, which will make sure that at least the
desktop launch is implemented.

If not implemented, the other two launch paths will behave as a normal
desktop launch.

This commit doesn't change any functionality yet, just puts the interface
in place.

[endlessm/eos-sdk#2681]