Refactor content loading out of constructor
Since we will need to take different actions after loading the content,
depending on which code path we are launched through, the
_load_all_content() method has to be refactored to take a callback,
instead of updating the UI itself when it is done loading.

Because of the way it works, it actually has to take two callbacks - one
which is called on error or when all content has been loaded, and another
"progress callback" which is called on each batch of results. (These are
Node-style callbacks and should be refactored to become either GIO-style
callbacks or Endless-style promises in a future revision.)

desktop_launch() now calls _load_all_content() and handles the UI updates
by passing in callbacks.

We already had two slightly different code paths calling
_load_all_content(), so this goes ahead and separates them. In addition
to desktop_launch(), content is also loaded when a new issue is populated
when UserSettingsModel:start-article changes. Only in the latter case do
we have to clean out existing content; previously this was lumped all
together.

[endlessm/eos-sdk#2681]