Show standalone page when opening search result
When navigating to an archived article via global search, the app should
open a standalone page with no navigation buttons on either side. This
implements the activate_search_result() method from EosKnowledge.Launcher
and checks whether the passed-in ID is in the archive or not. If it is,
then a standalone page is opened. If it is not, then the app simply
navigates to that page.

(Based on Rory's original code)

[endlessm/eos-sdk#2681]