Eliminate transparent gap while sliding in article
On template B, when an article slides in, the app background is visible
through a gap. This eliminates that gap, replacing it with a gray
background that matches the article background. Turns out WebKitWebView
was transparent by default, so before the article loaded, you could see
the app background through it.

Unfortunately, sometimes on first load, the webview flashes white
briefly. That's something we've seen on other apps like the programming
app, so I don't think it's a problem with this app specifically.

[endlessm/eos-sdk#2723]