Remove Fitness-specific styling from Reader CSS
The default Reader CSS should include just enough styling to show that
the styling is working, but not too much so that overrides in apps have
to override everything. For example, in the article snippet titles,
app overrides shouldn't have to undo italics by specifying font-weight:
normal.

This gives everything a reasonable size with the default font and shows
the three variations using bright primary colors.

[endlessm/eos-sdk#2851]