Add ContentObjectModel:original-uri
This property corresponds to the originalURI field in our content
databases. The source-uri property contains the URI from which the
content was downloaded during database build. This information isn't
really needed at runtime and only needs to be tracked for backwards
compatibility with old databases.

For Wikipedia and WikiHow articles, source-uri already contains the same
URI as original-uri would contain, because we download the articles
directly from their sources. However, for Reader apps, we preprocess the
articles and download them from the Pantheon server, so the two URIs are
not the same.

We should use original-uri to show the links back to the original article
for Wikipedia and WikiHow as well, but that is outside the scope of this
issue.

[endlessm/eos-sdk#2867]