Render attribution at bottom of Reader articles
This uses the new information that we get when entering feeds into a
Reader article in the Pantheon to render an attribution notice at the
bottom of Reader articles.

It includes a link back to the original blog, and if applicable, a link
to the license. All license names, and applicable license URIs, should be
internationalized.

[endlessm/eos-sdk#2867]