xapianQuery: use correct exact title clause in prefix search
We used a few separate code paths for forming exact title clauses
in xapianQuery. This introduced a bug in prefix search where we
would not capitalize the exact title clause in prefix search. Due
to a odd quirk of xapian, this means the query would not be stemmed
if the user capitalized their query, and would be stemmed if it was
lowercase.

Using the function in all code paths should avoid these errors
[endlessm/eos-sdk#2771]