articleObjectModel: Load HTML from disk
Our new databases have the HTML separated out from the Xapian database
so that we can read the metadata during search without loading the
entire blob, which will hopefully minimize seek times a lot more.

To keep compatibility, we insert a fancy new "html" getter which will
properly return the articleBody contents if it exists, and fall back
to reading the content URI if it does not.