Display webview tooltips on link elements
When hovering over a link element in the text, a tooltip pops up,
pointing to the link element and displaying the URI.

We get the area of the link element, which is known only from the DOM, by
querying a custom web extension over DBus.

The readerPresenterSmokeTest.js test demonstrates this functionality;
hover over the link on the front page to see where it goes.

(Philip & Fernando)

[endlessm/eos-sdk#2744]