Add disabled incomplete test for web extension
Due to a bug in Jasmine 2.0, async beforeEach setup functions do not
work correctly, and this test is not possible. This sketches a possible
implementation, but we need to upgrade to Jasmine GJS for it to work;
therefore we will not spend any additional time on it right now.

(Philip & Fernando)

[endlessm/eos-sdk#2744]