engine: don't test exact query clause structure
We already test the structure in the xapian_query_lib tests (and
even a little too strictly over there I think). The engine model
isn't directly responsible for forming all the ANDs and
exact_titles: and what not, so we shouldn't test for those in
an engine unit test.

Softened the tests to just look for general brokenness in the query
clause without asserting the entire structure.
[endlessm/eos-sdk#2771]