presenter: use incremental search
We will use incremental search for the search results page in addition
to the autocomplete results. It gives a consistent experience when
activating from global search, and is a poor mans substitute for
a "did you mean?" search feature.

We also will also stick to title only search, as that gives better
results.
[endlessm/eos-sdk#2771]