Take height-for-width/width-for-height into account
SpaceContainer now requests the correct size for HFW/WFH widgets, such as
labels.

Due to time constraints, this implementation isn't complete -- in
vertical orientation, the total width request of the SpaceContainer won't
change according to its height. It will just be equal to the minimum of
all the children, as it was before. This works for our current use cases.

[endlessm/eos-sdk#2882]