Fix navigation to in-issue article.
When going back to an "in-issue" article from an "archived" article, we were
erroneously using the model.article_number to navigate to the article_models
array, avoiding the app to reach the actual page that was wanted.

We now filter the article_models array with the desired article_number to
navigate to the correct index into the array.

[endlessm/eos-sdk#2743]