Override image credits for non-Reader apps
Currently we don't have attribution information for images in non-Reader
apps. Rather than special-casing to deal with missing files, we just set
the attribution file to be null for non-Reader apps. (It's lucky the
divide between haves and have-nots falls along template lines.)

Revert this commit when we add the missing information to
eos-knowledge-apps.

[endlessm/eos-sdk#2939]