scss: add unicode to all toplevels to force charset
I have no idea why this is needed TBH. What sass is doing is
perfectly valid. There should be no need to specify utf8
if the content of an css file is all plain ascii. There should
be no need in the first place, as our html should be all
interpreted as utf8, which in turn should imply that the
css is interpreted as utf8 by default.

But for some reason webkit is not displaying our css is if does
not have the charset marker. See the issue for more discussion.
[endlessm/eos-sdk#2980]