article.mst: add charset declaration to html
Because of the way load_html is handled, this doesn't do squat,
and the page will actually be interpreted as latin1, or utf-16.

But its good practice to specify the charset and utf-8 is what
we would want. So in a bright sunny future where this property
is actually respected by webkit, we should have this in.
[endlessm/eos-sdk#3026]