Don't directly load article html, load through ekn uri handler
This saves us from a bug where the directly loaded html was
sometimes converted to utf-16 within webkit, which causing webkit
to think every single css and javascript asset we used was also
encoded the same.

By just leveraging our uri handler to load articles as well, we
can make sure the page and all scripts and css are treated as
utf-8 encoded.
[endlessm/eos-sdk#2961]