Improve validation of EKN ID
EKN ID is a URI with the schema "ekn", an optional path component of
"api", followed by a domain and a hash. The domain may be anything that
is valid in an application ID when prefixed by a reverse internet domain
name. The hash is exactly 16 hex digits.

Refactor the tests so that the EKN ID validation function is tested
separately.

[endlessm/eos-sdk#3048]