webkitURIHandlers: squash bug when loading images for old article
We were hitting a segfault when finishing async uri requests for
images in an article that we are no longer viewing.

To be honest not sure we have the full picture here, rushing this
out. But it seems to fix things to ignore uri request for a webview
that has been destroyed or has since navigated to a new page.
[endlessm/eos-sdk#3078]