Makefile: consolidate .am.inc files back into main Makefile
We had our Makefile distributed all over a bunch of directories.
The idea was to keep the modular nature of recursive make Makefiles
without the many downside of recursive make.

However, as this scaled it became less and less feasible, because
all the make variables across 7 different files were all scoped to
the same namespace.

So we'll just bring everything back into the main Makefile so its
easier to tell when a variable is declared and when its appended
to/modified.
[endlessm/eos-sdk#3099]