Add EKN_VERSION
App bundles should provide an EKN_VERSION file in their toplevel ekn/data
directory. It contains a single integer which indicates the bundle data format
version; if the file's not there (as in the case of all legacy bundles), it's
assumed to be 1.

Anytime the bundle data changes in a way that necessitates a change in
eos-knowledge-lib, the EKN_VERSION value should be incremented. This way,
knowledge lib code needs to only check the version number to support backwards
compatibility.

[endlessm/eos-sdk#2917]