queryObject: handle all query parser string construction
A redefinition of roles. Rather than having xapianQuery define a
whole mess of functions which the engine call into in a series of
if statements, we'll go for a cleaner interface.

The query object will house all the query construction code, and form
the entire query in a single method that the engine can call into.

Also moved cutoff values, xapian value numbers and other xapian
query details into queryObject so they are all in one place.
[endlessm/eos-sdk#3100]