Move in encyclopedia code and assets
Deprecating the eos-encyclopedia repo. The encyclopedia app is
essentially just another template type for a knowledge app. The
code should live with the other template types here in knowledge
lib. This will save us the headache of needing to simultaneously
patch this repo and encyclopedia's when we make a change to common
javascript api.

This commit attempt to just move in the code, touching as little
as possible while getting things working under knowledge lib. The
only things changed from how they were in eos-encyclopedia should
be:
 - filenames
 - imports paths
 - gresource uris
 - webkit extension registering
 - gettext setup
Other than that, code should be verbatim from the encyclopedia
repo.
[endlessm/eos-sdk#3102]