Redo debian packaging for private C libraries
We switched up our knowledge lib structure so that all our
javascript and C are private to our package, and only set up for
use by our bin scripts.

This will change up the packing so that instead of the packaging
split for a library with a dev and gir package, we just have one
eos-knowledge-lib package with all our bin scripts and code.

We keep a doc library for our private C lib documentation, for our
own convenience.

Switched up runtime dependencies as so we no longer use the
gir:Depends. I think that was always pretty broken. There's no
guarantee that the gir includes we need for our C library has a 1:1
mapping with the libraries actually used by all our javascript code.
There were a number of needless includes in the gir files, a bunch
of missing includes for the debian, so think explicit runtime deps
is a better way to go.
[endlessm/eos-sdk#3108]