configure: rename package eos-knowledge -> eos-knowledge-lib
To agree with our new debian package names, which in turn are
agreeing with the new scheme naming schemes we have for things.
Since we were renaming everything anyway for debian packaging,
seemed like a good time to do this.

After this, the only things that are still names eos-knowledge-lib
are the repo and the debian source package. These could both be
changed, but would have to touch a bunch more external things
(the jenkins job, repo lists in eos-build...) so for know those
will stay as is, and it hopefully won't affect things.
[endlessm/eos-sdk#3108]