Rename searchUtils.js utils.js
I suspect the reason this was originally named utils was collisions
in the imports object. Now that we have a common root to all our
imports it is no longer necessary. We were importing it as Utils
in code anyway, and imports.search.utils is a lot less redundant
than imports.search.searchUtils.
[endlessm/eos-sdk#3108]