smoke-tests: Fix smoke test imports
To work with the new location for the js code and the new name of
the EosKnowledgePrivate library.

The EosKnowledgePrivate library will sometimes be imported but not
used! This is to get the resource in the library properly registered.

Uses a smoke-test-runner to set up all paths correctly. This does
not currently support out of tree builds, but don't think that will
be an issue as the smoke tests are only intended for local development.
[endlessm/eos-sdk#3108]