gresource: Move out of private C library
The gresource will now be a regular just a regular file installed
into pkgdatadir.

We will register it right before creating our presenter, and in
any tests that actually need it. Eventually we should implement a
better file mocking solution for our unit tests, so none of them
to need register the actual production gresource.

This means a couple things...
 * We can drop our overrides hack forcing the C library to
   initialize on import from gjs, and instead have it just initialize
   when its used like every other import. And no more unused
   imports.gi.EosKnowledgePrivate just for its side effects.
 * Now that the resource is outside of the lib, we can register
   it with the webkit process as well. (in a coming commit)
[endlessm/eos-sdk#3004]