testEngine: move all test content into js file
We had a few bits of json we were reading from disk. Did not
make sense to have that split out, as some longer bits of json
are contained inline in the file.

It was implemented buggily where the files were loaded outside of
a before each, but the objects were actually modified during tests
inside mock_engine_request. This could create a weird race
condition where our json was double "stringified."

Moved everything into the actual test file, and simplified the
mock_engine_request function so it does not modify the objects passed
in.
[endlessm/eos-sdk#2855]