Always show 'show-more' button when we have >6 categories
Previously we would only show the 'show-more' button
when the space container on the home page could
not show all its children. However, now we cap
the number of children we give to the space container
to 6, so it always thinks it can show all its children.
Hence the show-more categories button is never visible.
This fixes it by checking if we have extra categories
to show in pack_cards and setting a flag there
to show the button.

[endlessm/eos-sdk#3214]